% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize a Lineage into Nearest Supplied Parental Lineages}
\usage{
categorize(
  graph_obj = NULL,
  lineage_of_interest = NULL,
  parental_group = NULL,
  default_lineage = NA_character_
)
}
\arguments{
\item{graph_obj}{an igraph object constructed from \code{read_lineages()}
depicting all SARS-CoV-2 lineage parent/child relationships.}

\item{lineage_of_interest}{A character string depicting a SARS-CoV-2 lineage
to be collapsed.}

\item{parental_group}{A character vector of a lineage(s) which we would like
to categorize into.}

\item{default_lineage}{If the lineage is not grouped, what value would you
like to assign to the lineage (default: NA)?}
}
\value{
A character vector of SARS-CoV-2 lineages
}
\description{
\code{categorize()} returns the nearest SARS-CoV-2 lineage relative to lineage of
interest.
}
\examples{
#An example of identifying the nearest parental lineage to "KP.3.1.1"
lineages <- read_lineages()

categorize(graph_obj = lineages, lineage_of_interest = "KP.3.1.1",
 parental_group =  c("KP.1", "KP.2", "KP.3") )

}
