% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_binning.R
\name{cumsum_binning}
\alias{cumsum_binning}
\title{Empirically Determine Lineages Groups Which Meet a Certain Threshold}
\usage{
cumsum_binning(graph_obj, lineages, values, threshold)
}
\arguments{
\item{graph_obj}{an igraph object constructed from \code{read_lineages()}
depicting all SARS-CoV-2 lineage parent/child relationships.}

\item{lineages}{A vector of observed SARS-CoV-2 lineages.}

\item{values}{A vector of observed SARS-CoV-2 lineage quantities (numeric).}

\item{threshold}{Minimum numeric quantity required to identify lineages (numeric).
Must be greater than 0}

\item{parental_group}{A character vector of a lineage(s) which we would like
to categorize into.}
}
\value{
A character vector of SARS-CoV-2 lineages
}
\description{
\code{cumsum_binning()} returns SARS-CoV-2 lineages which can be used to as groups
for \code{categorize()}.
}
\examples{
#An example of identifying the nearest parental lineage to "KP.3.1.1"
lineage_graph <- read_lineages()
lineages <- c("KP.3", "KP.3.1", "KP.3.1.1", "KP.3.1.4", "XEF" )
values = c( 1, 1, 5, 5, 10) 

# Identify lineages which can have sub-lineages collapsed to meet a count of 10
cumsum_binning(lineage_graph, lineages, values, 10)

}
