% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{decedents}
\alias{decedents}
\title{Lists All Sub-lineages of a Given Lineage}
\usage{
decedents(
  graph_obj = NULL,
  lineage_of_interest = NULL,
  remove_recombinants = FALSE
)
}
\arguments{
\item{graph_obj}{an igraph object constructed from \code{read_lineages()}
depicting all SARS-CoV-2 lineage parent/child relationships.}

\item{lineage_of_interest}{A character string depicting a SARS-CoV-2 lineage.}

\item{remove_recombinants}{Logical value to specify removal of recombination}
}
\value{
A character vector of SARS-CoV-2 lineages
}
\description{
\code{decedents()} returns all SARS-CoV-2 lineages descending the lineage of
interest.
}
\examples{
#An example of pulling all SARS-CoV-2 sub-lineages of KP.3.1.1, while removing
#all recombinants and their decedents.
lineages <- read_lineages()
decedents(graph_obj = lineages, lineage_of_interest = "KP.3.1.1", remove_recombinants = TRUE)

#An example of pulling all sub-lineages of XFG.
decedents(graph_obj = lineages, lineage_of_interest = "XFG")
}
