% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect_groups}
\alias{detect_groups}
\title{Empirically Determine Predominant Lineage Groups}
\usage{
detect_groups(graph_obj, lineages, values, threshold = NULL, n = NULL)
}
\arguments{
\item{graph_obj}{an igraph object constructed from \code{read_lineages()}
depicting all SARS-CoV-2 lineage parent/child relationships.}

\item{lineages}{A vector of observed SARS-CoV-2 lineages.}

\item{values}{A vector of observed SARS-CoV-2 lineage quantities (numeric).}

\item{threshold}{Minimum numeric quantity required to identify lineages (numeric).
Must be greater than 0. Predominant lineage groups are identified via cumulative
sum of observed lineages with reset after a threshold is achieved.}

\item{n}{Number of predominant lineages which capture all observed sub-lineage frequencies.
Top n lineage groups are identified via bottom up pruning.}
}
\value{
A character vector of SARS-CoV-2 lineages
}
\description{
\code{detect_groups()} returns SARS-CoV-2 lineages which can be used to as groups
for \code{categorize()}.
}
\examples{
# Identify predominant lineages which, when pooled with their sub-lineages, meet a count of 10
lineage_graph <- read_lineages()
lineages <- c("KP.3", "KP.3.1", "KP.3.1.1", "KP.3.1.4", "XEF" )
values = c( 1, 1, 5, 5, 10) 

detect_groups(lineage_graph, lineages, values, threshold = 10)

#Identify the top 2 lineages which capture all observed sub-lineages
detect_groups(lineage_graph, lineages, values, n = 2)


}
