% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_binning.R
\name{.annotate}
\alias{.annotate}
\title{Assign a numeric attribute to vertices in network}
\usage{
.annotate(graph_obj, lineages, values, default = 0)
}
\description{
Internal function used by \code{cumsum_binning()}.
}
\keyword{internal}
