% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{.assign_rcbnt}
\alias{.assign_rcbnt}
\title{Assign recombination status}
\usage{
.assign_rcbnt(graph_obj, parental_group, closest_parents, lineage_of_interest)
}
\description{
Internal function used by \code{categorize()}.
}
\keyword{internal}
