% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{.closest_parent}
\alias{.closest_parent}
\title{Identify closest parental group for a given lineage of interest}
\usage{
.closest_parent(group_indices, lineage_indices, graph_obj)
}
\description{
Internal function used by \code{categorize()}.
}
\keyword{internal}
