% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{.split_recombinants}
\alias{.split_recombinants}
\title{Internal function used by \code{detect_groups()}.}
\usage{
.split_recombinants(graph_obj)
}
\description{
Internal function used by \code{detect_groups()}.
}
\keyword{internal}
