% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_binning.R
\name{.subset_graph}
\alias{.subset_graph}
\title{Subsets network to observed lineages}
\usage{
.subset_graph(graph_obj, root, lineages)
}
\description{
Internal function used by \code{cumsum_binning()}.
}
\keyword{internal}
