% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_binning.R
\name{.threshold_cumsum}
\alias{.threshold_cumsum}
\title{Internal function used by \code{cumsum_binning()}.}
\usage{
.threshold_cumsum(graph_obj, threshold)
}
\description{
Internal function used by \code{cumsum_binning()}.
}
\keyword{internal}
