% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{parents}
\alias{parents}
\title{Lists All Direct Parents to a Given Lineage}
\usage{
parents(graph_obj = NULL, lineage_of_interest = NULL)
}
\arguments{
\item{graph_obj}{an igraph object constructed from \code{read_lineages()}
depicting all SARS-CoV-2 lineage parent/child relationships.}

\item{lineage_of_interest}{A character string depicting a SARS-CoV-2 lineage.}
}
\value{
A character vector of SARS-CoV-2 lineages
}
\description{
\code{parents()} returns all direct SARS-CoV-2 parental lineages to the lineage of
interest.
}
\examples{
#An example of pulling all direct parent(s) of KP.3.1.1.
lineages <- read_lineages()
parents(graph_obj = lineages, lineage_of_interest = "KP.3.1.1")

#An example of pulling all direct parents of XFG.
parents(graph_obj = lineages, lineage_of_interest = "XFG")

}
