% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lineages.R
\name{read_lineages}
\alias{read_lineages}
\title{Reads and Converts SARS-CoV-2 Lineages into an igraph object}
\usage{
read_lineages()
}
\value{
An igraph object.
}
\description{
\code{read_lineages()} downloads the current SARS-CoV-2 PANGO lineage data from
outbreak.info, converts it to a tabular structure, and returns a directed
network
}
\examples{
lineage_network <- read_lineages()

}
