% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_binning.R
\name{.split_recombinants}
\alias{.split_recombinants}
\title{Internal function used by \code{cumsum_binning()}.}
\usage{
.split_recombinants(graph_obj)
}
\description{
Internal function used by \code{cumsum_binning()}.
}
\keyword{internal}
