% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{.threshold_cumsum}
\alias{.threshold_cumsum}
\title{Internal function used by \code{detect_groups()}.}
\usage{
.threshold_cumsum(graph_obj, threshold)
}
\description{
Internal function used by \code{detect_groups()}.
}
\keyword{internal}
