% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lineages.R
\name{read_lineages}
\alias{read_lineages}
\title{Reads in SARS-CoV-2 Lineage Data}
\usage{
read_lineages()
}
\value{
An igraph object.
}
\description{
\code{read_lineages()} reads in the current SARS-CoV-2 PANGO lineage data from
outbreak.info and represents this information in a directed network
}
\examples{
lineage_network <- read_lineages()

}
